code cleanup
Former-commit-id: b1b04f013e7eda4a1cad5ee907b2f5c58da9f4c8

+++ src/ves/xplorer/scenegraph/rtt/Unit.cxx
@@ -191,7 +191,7 @@ const Unit::TextureMap& Unit::GetInputTextureMap() const
     return mInputTextures;
 }
 ////////////////////////////////////////////////////////////////////////////////
-osg::Texture* Unit::GetOutputTexture( int mrt ) const
+osg::Texture* Unit::GetOutputTexture( size_t mrt ) const
 {
     if( mrt < mOutputTextures.size() )
     {
